pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "auto-capture"
version = "0.1.0"
description = "macOS auto-screenshot tool — captures on mouse click with annotation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "LaunchDock Team" },
]
keywords = ["screenshot", "macos", "automation", "capture"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
]

dependencies = [
    "pyobjc-core>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "Pillow>=10.0",
    "tomli>=2.0; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
auto-capture = "auto_capture.cli:main"

[tool.setuptools.packages.find]
include = ["auto_capture*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
